+++ engine/src/test/java/org/archive/crawler/processor/BrowserProcessorTest.java
@@ -8,11 +8,15 @@ import org.archive.modules.fetcher.FetchHTTP2;
 import org.archive.net.UURIFactory;
 import org.archive.url.URIException;
 import org.archive.util.Recorder;
+import org.eclipse.jetty.proxy.ProxyHandler;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.ServerConnector;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 import org.junit.jupiter.api.io.TempDir;
 
 import java.io.IOException;
+import java.net.Inet4Address;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
 import java.nio.file.Path;
@@ -65,6 +69,33 @@ class BrowserProcessorTest {
         assertFalse(crawlURI.getAnnotations().contains("browser"), "navigation should have aborted");
     }
 
+    @Test
+    public void testHttpProxy() throws Exception {
+        InetAddress localhost = Inet4Address.getLoopbackAddress();
+        Server proxyServer = new Server(new InetSocketAddress(localhost, 0));
+        proxyServer.setHandler(new ProxyHandler.Forward());
+        proxyServer.start();
+        try {
+            fetcher.setHttpProxyHost(localhost.getHostAddress());
+            fetcher.setHttpProxyPort(((ServerConnector)proxyServer.getConnectors()[0]).getLocalPort());
+
+            CrawlURI crawlURI = newCrawlURI(baseUrl);
+            fetcher.process(crawlURI);
+            assertEquals(200, crawlURI.getFetchStatus());
+            browserProcessor.innerProcess(crawlURI);
+
+            var outLinks = new ArrayList<>(crawlURI.getOutLinks());
+            assertEquals("/link", outLinks.get(0).getUURI().getPath());
+            assertTrue(crawlURI.getAnnotations().contains("browser"));
+            assertEquals("true", crawlURI.getHttpResponseHeader("Used-Proxy"));
+            assertEquals("true", subrequests.get(0).getHttpResponseHeader("Used-Proxy"));
+
+            logger.log(DEBUG, "Subrequests: {0}", subrequests);
+        } finally {
+            proxyServer.stop();
+        }
+    }
+
     private CrawlURI newCrawlURI(String uri) throws URIException {
         CrawlURI curi = new CrawlURI(UURIFactory.getInstance(uri));
         Recorder recorder = new Recorder(tempDir.toFile(), "fetcher");
@@ -113,6 +144,9 @@ class BrowserProcessorTest {
                     exchange.getResponseHeaders().add("Content-Disposition", "attachment; filename=heritrix-test.bin");
                 }
                 default -> status = 404;
+            }
+            if (exchange.getRequestHeaders().containsKey("Via")) {
+                exchange.getResponseHeaders().add("Used-Proxy", "true");
             }
             exchange.getResponseHeaders().add("Content-Type", contentType);
             exchange.sendResponseHeaders(status, 0);